整合示例到一个页面中
